move storage types to pkg storage
replace repetitive code with appropriate function
