+++ pkg/virt-controller/services/rendervolumes.go
@@ -14,8 +14,8 @@ import (
 	"kubevirt.io/kubevirt/pkg/config"
 	"kubevirt.io/kubevirt/pkg/hooks"
 	hostdisk "kubevirt.io/kubevirt/pkg/host-disk"
+	"kubevirt.io/kubevirt/pkg/storage/types"
 	"kubevirt.io/kubevirt/pkg/util"
-	"kubevirt.io/kubevirt/pkg/util/types"
 )
 
 type VolumeRendererOption func(renderer *VolumeRenderer) error
