+++ pkg/virt-controller/watch/BUILD.bazel
@@ -28,6 +28,7 @@ go_library(
         "//pkg/monitoring/vmistats:go_default_library",
         "//pkg/service:go_default_library",
         "//pkg/storage/snapshot:go_default_library",
+        "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/cluster:go_default_library",
         "//pkg/util/lookup:go_default_library",
