+++ pkg/virt-controller/watch/export/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "//pkg/certificates/triple:go_default_library",
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/storage/types:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/snapshot:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
