+++ pkg/virt-handler/hotplug-disk/BUILD.bazel
@@ -14,6 +14,8 @@ go_library(
         "//pkg/hotplug-disk:go_default_library",
         "//pkg/safepath:go_default_library",
         "//pkg/unsafepath:go_default_library",
+        "//pkg/storage/types:go_default_library",
+        "//pkg/util:go_default_library",
         "//pkg/virt-handler/cgroup:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-handler/virt-chroot:go_default_library",
